backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
fal-client==0.5.6
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4
google-cloud-storage==2.18.2
python-dotenv==1.0.1
aiofiles==24.1.0
websockets==13.1
httpx==0.27.2
